Close #17; Add Ounit integration test
This commit adds an OUnit integration test to verify the correctness of the AST for a basic expression. The test suite file is the main executable entry point that runs all the tests. The arithmetic parser test file tests the ast of arithmetic expressions.

To execute the tests, navigate to either the root project directory or the test directory and run: .

The arithmetic parser test file should ideally be placed inside the arithmetic folder, but I encountered difficulties configuring Dune to locate it

One difficulty was to make the compiler modules accessible from outside the src folder. To achieve this the dune file inside the src folder was updated to make a library that includes the AST, Parser, and Lexer modules. This library is then imported into the Dune file within the tests folder