Add organization and control flow parser test
This commit tries to reorginize the tests folder to minimize future clutter. Each folder inside the tests directory is used to test a specific area, in this case we have an arithmetic and control flow folders where related tests are placed. To make the tests accessible to the parent tests directory you have to use the dune file to make it into a library. Whenever you make new folders remember to copy the dune file over, change the name of the library, and make a suite that contains all the tests inside the folder. The file containing this suite should be named something unique so the test suites can be run without conflicts in the root directory