Add Ounit integration test (#40)
* Close #17; Add Ounit integration test

This commit adds an OUnit integration test to verify the correctness of the AST for a basic expression. The test suite file is the main executable entry point that runs all the tests. The arithmetic parser test file tests the ast of arithmetic expressions.

To execute the tests, navigate to either the root project directory or the test directory and run: .

The arithmetic parser test file should ideally be placed inside the arithmetic folder, but I encountered difficulties configuring Dune to locate it

One difficulty was to make the compiler modules accessible from outside the src folder. To achieve this the dune file inside the src folder was updated to make a library that includes the AST, Parser, and Lexer modules. This library is then imported into the Dune file within the tests folder

* Add organization and control flow parser test

This commit tries to reorginize the tests folder to minimize future clutter. Each folder inside the tests directory is used to test a specific area, in this case we have an arithmetic and control flow folders where related tests are placed. To make the tests accessible to the parent tests directory you have to use the dune file to make it into a library. Whenever you make new folders remember to copy the dune file over, change the name of the library, and make a suite that contains all the tests inside the folder. The file containing this suite should be named something unique so the test suites can be run without conflicts in the root directory